test(developer): shuffle tests
- split out test-strs from test-compiler-e2e
- split out test-linter from test-compiler-e2e
